requirements.txt
mido
python-rtmidi
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
pyyaml
spidev
Pillow
numpy #sudo apt-get install libatlas-base-dev